clustering/requirements.txt
numpy
matplotlib
argparse
tqdm
Pillow
